Rename tool to CreateBaselineUpdatePR
