+++ src/SourceBuild/content/eng/tools/CreateBaselineUpdatePR/Pipelines.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace PrBaselinePublisher;
+namespace CreateBaselineUpdatePR;
 
 public enum Pipelines
 {
